esercizio 2 lezione 5 per casa  terminato

+++ Lezione4/printing_models.py
@@ -8,12 +8,12 @@ favorite_book("Il nome del vento")
 
 
 
-def somma(a:int,b:int)->int:
+def somma(a:float,b:float)->float:
     return a+b
 
 
 print(somma(10,5))
 
 
-def divisione(a:int,b:int)->int:
+def divisione(a:float,b:float)->float:
     return a/b
